+++ services/api/src/owl/db/gen_table.py
@@ -1036,7 +1036,12 @@ class GenerativeTable:
                 limit = total - offset
             if order_descending:
                 offset = max(0, total - limit - offset)
-            rows = table._dataset.to_table(offset=offset, limit=limit).to_pylist()
+            if columns is not None:
+                if "ID" not in columns:
+                    columns.insert(0, "ID")
+                if "Updated at" not in columns:
+                    columns.insert(1, "Updated at")
+            rows = table._dataset.to_table(columns=columns, offset=offset, limit=limit).to_pylist()
             rows = sorted(rows, reverse=order_descending, key=lambda r: r["ID"])
             rows = self._post_process_rows(
                 rows,
@@ -1170,7 +1175,7 @@ class GenerativeTable:
     def export_csv(
         self,
         table_id: TableName,
-        columns: list[ColName] | None = None,
+        columns: list[ColName],
         file_path: str = "",
         delimiter: CSVDelimiter | str = ",",
     ) -> pd.DataFrame: